+++ packages/plugin-dev-essentials/src/services/test-service.ts
@@ -288,7 +288,9 @@ export class TestService {
 
     try {
       await this.ensureDaemons();
-      await this.buildAndPublishSnapshot()
+      if (cliSource === 'cliRegistry') {
+        await this.buildAndPublishSnapshot()
+      }
     } catch (err) {
       return {
         success: false as const,
@@ -306,7 +308,7 @@ export class TestService {
       }
     );
 
-    // context.logger.info(`RUNNING STAGE0 TESTS - ${cliSource}`);
+    context.logger.info(`RUNNING STAGE0 TESTS - ${cliSource}`);
 
     const stage0Res = await packageManagerService.run({
       command: `jest 'stage0'${